requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
oauthlib==3.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
typing_extensions==4.12.2
urllib3==2.2.2
